+++ clang/test/dpct/test_api_level/Util/api_test21.cu
@@ -3,13 +3,13 @@
 // RUN: FileCheck --input-file %T/Util/api_test21_out/count.txt --match-full-lines %s
 // RUN: rm -rf %T/Util/api_test21_out
 
-// CHECK: 2
+// CHECK: 1
 // TEST_FEATURE: Util_conj
 
 #include <cuComplex.h>
 
 __device__ void foo1() {
-    double2 a;
+    float2 a;
 
     auto c = cuConj(a);
